Use lambda as parallel_body in old OpenCV
