+++ ov_core/src/track/Grider_FAST.h
@@ -31,25 +31,9 @@
 #include <opencv2/highgui/highgui.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
 
+#include "../utils/lambda_body.h"
 
 namespace ov_core {
-
-    /**
-     * This is a utility class required to build with older version of opencv
-     * On newer versions this doesn't seem to be needed, but here we just use it to ensure we can work for more opencv version.
-     * https://answers.opencv.org/question/65800/how-to-use-lambda-as-a-parameter-to-parallel_for_/?answer=130691#post-id-130691
-     */
-    class LambdaBody : public cv::ParallelLoopBody {    
-    public:
-        explicit LambdaBody(const std::function<void(const cv::Range &)> &body) {
-            _body = body;
-        }
-        void operator() (const cv::Range & range) const override {
-            _body(range);
-        }    
-    private:
-        std::function<void(const cv::Range &)> _body;
-    };
 
     /**
      * @brief Extracts FAST features in a grid pattern.
